DROWSY/DROWSY/requirements.txt
pandas==2.1.0
numpy==1.24.3
matplotlib==3.7.2
scipy==1.11.2
scikit-learn>=1.0.0
joblib>=1.3.0
seaborn>=0.12.0
pygame>=2.5.0
spotipy>=2.23.0
python-dotenv>=1.0.0
